+++ user/ble_comm.c
@@ -656,7 +656,8 @@ extern void MbsTransShowBuffInfo();
 // Event: gecko_evt_system_external_signal_id
 //
 //**********************************************************************************************
-uint32 extern ShowSqueNum();
+extern uint32 ShowSqueNum();
+extern void TestIvUpdate();
 uint32 EvtSysExternalSignalProc(PCmdPacket pEvent)
 {
     uint32 ret_code=TRUE;
@@ -679,6 +680,7 @@ uint32 EvtSysExternalSignalProc(PCmdPacket pEvent)
     	dprint("--- PB1 ON\r\n");
 #ifdef BTM_TRANSMITTER
     	//for test
+    	//TestIvUpdate();
     	//SetNodeStatus(NS_IVI_UPDATE,ON);
     	/*result=Cmd_mt_set_element_seqnum(PRIMARY_ELEM, 0xffffff)->result; //set sequence num = 0
     	dprint("Cmd_mt_set_element_seqnum:result:0x%X\r\n",result);*/